// part9/patientor/front-end/src/types.ts
export interface Diagnosis {
    code: string;
    name: string;
    latin?: string;
}

export enum Gender {
    Male = "male",
    Female = "female",
    Other = "other",
}

export interface Patient {
    id: string;
    name: string;
    occupation: string;
    gender: Gender;
    ssn?: string;
    dateOfBirth?: string;
    entries: Entry[];
}

interface EntryBase {
    id: string;
    date: string;
    specialist: string;
    diagnosisCodes?: Array<Diagnosis["code"]>;
    description: string;
}

export interface HospitalEntry extends EntryBase {
    type: "Hospital";
    discharge: { date: string; criteria: string };
}

export interface OccupationalHealthcareEntry extends EntryBase {
    type: "OccupationalHealthcare";
    employerName: string;
    sickLeave?: { startDate: string; endDate: string };
}

export interface HealthCheckEntry extends EntryBase {
    type: "HealthCheck";
    healthCheckRating: number;
}

export enum HealthCheckRating {
    "Healthy" = 0,
    "LowRisk" = 1,
    "HighRisk" = 2,
    "CriticalRisk" = 3,
}

export type Entry =
    | HospitalEntry
    | OccupationalHealthcareEntry
    | HealthCheckEntry;